Refactor passes/callexpr/helper and passes/selectorexpr/helper into passes/helper
Less indirection to find what you need.
