+++ passes/R001/R001.go
@@ -8,7 +8,7 @@ import (
 	"golang.org/x/tools/go/analysis"
 
 	"github.com/bflad/tfproviderlint/passes/commentignore"
-	"github.com/bflad/tfproviderlint/passes/callexpr/helper/schema/resourcedataset"
+	"github.com/bflad/tfproviderlint/passes/helper/schema/resourcedatasetcallexpr"
 )
 
 const Doc = `check for ResourceData.Set() calls using complex key argument
