+++ passes/V002/V002.go
@@ -3,7 +3,7 @@ package V002
 import (
 	"github.com/bflad/tfproviderlint/helper/analysisutils"
 	"github.com/bflad/tfproviderlint/helper/terraformtype/helper/validation"
-	"github.com/bflad/tfproviderlint/passes/selectorexpr/helper/validation/cidrnetwork"
+	"github.com/bflad/tfproviderlint/passes/helper/validation/cidrnetworkselectorexpr"
 )
 
 var Analyzer = analysisutils.DeprecatedWithReplacementSelectorExprAnalyzer(
