+++ passes/V004/V004.go
@@ -3,7 +3,7 @@ package V004
 import (
 	"github.com/bflad/tfproviderlint/helper/analysisutils"
 	"github.com/bflad/tfproviderlint/helper/terraformtype/helper/validation"
-	"github.com/bflad/tfproviderlint/passes/selectorexpr/helper/validation/singleip"
+	"github.com/bflad/tfproviderlint/passes/helper/validation/singleipselectorexpr"
 )
 
 var Analyzer = analysisutils.DeprecatedWithReplacementSelectorExprAnalyzer(
