+++ passes/V005/V005.go
@@ -3,7 +3,7 @@ package V005
 import (
 	"github.com/bflad/tfproviderlint/helper/analysisutils"
 	"github.com/bflad/tfproviderlint/helper/terraformtype/helper/validation"
-	"github.com/bflad/tfproviderlint/passes/selectorexpr/helper/validation/validatejsonstring"
+	"github.com/bflad/tfproviderlint/passes/helper/validation/validatejsonstringselectorexpr"
 )
 
 var Analyzer = analysisutils.DeprecatedWithReplacementSelectorExprAnalyzer(
