+++ passes/helper/validation/validateregexpselectorexpr/validateregexpselectorexpr.go
