+++ passes/helper/validation/validateregexpselectorexpr/validateregexpselectorexpr_test.go
